first working version

+++ src/data_gradients/feature_extractors/common/mean_std.py
@@ -17,8 +17,8 @@ class MeanAndSTD(FeatureExtractorAbstract):
 
     def update(self, data: BatchData):
         for image in data.images:
-            self._hist[data.split]["mean"].append(torch.mean(image, [1, 2]))
-            self._hist[data.split]["std"].append(torch.std(image, [1, 2]))
+            self._hist[data.split]["mean"].append(torch.mean(image.float(), [1, 2]))
+            self._hist[data.split]["std"].append(torch.std(image.float(), [1, 2]))
 
     def _aggregate(self, split: str):
